refactor: move `getUserRoleFromHistory` into commons
