+++ packages/gateway/src/features/registration/type-resolvers.ts
@@ -9,10 +9,7 @@
  * Copyright (C) The OpenCRVS Authors located at https://github.com/opencrvs/opencrvs-core/blob/master/AUTHORS.
  */
 
-import {
-  getPresignedUrlFromUri,
-  getUserRoleFromHistory
-} from '@gateway/features/registration/utils'
+import { getPresignedUrlFromUri } from '@gateway/features/registration/utils'
 import { getSignatureExtension } from '@gateway/features/user/type-resolvers'
 
 import {
@@ -86,7 +83,8 @@ import {
   isTaskOrTaskHistory,
   resourceIdentifierToUUID,
   Address,
-  notCorrectedHistory
+  notCorrectedHistory,
+  getUserRoleFromHistory
 } from '@opencrvs/commons/types'
 
 import { GQLQuestionnaireQuestion, GQLResolver } from '@gateway/graphql/schema'
