+++ packages/gateway/src/features/registration/utils.ts
@@ -17,7 +17,6 @@ import {
   GQLDeathRegistrationInput,
   GQLMarriageRegistrationInput
 } from '@gateway/graphql/schema'
-import { PractitionerRole, Resource } from '@opencrvs/commons/types'
 
 export async function getPresignedUrlFromUri(
   fileUri: string,
@@ -64,42 +63,4 @@ export async function setCollectorForPrintInAdvance(
   })
 
   return details
-}
-
-export type IPractitionerRoleHistory = (Resource & PractitionerRole)[]
-export const getUserRoleFromHistory = (
-  practitionerRoleHistory: IPractitionerRoleHistory,
-  lastModified: string
-) => {
-  const practitionerRoleHistorySorted = practitionerRoleHistory.sort((a, b) => {
-    if (a.meta?.lastUpdated === b.meta?.lastUpdated) {
-      return 0
-    }
-    if (a.meta?.lastUpdated === undefined) {
-      return 1
-    }
-    if (b.meta?.lastUpdated === undefined) {
-      return -1
-    }
-    return (
-      new Date(b.meta?.lastUpdated).valueOf() -
-      new Date(a.meta?.lastUpdated).valueOf()
-    )
-  })
-
-  const result = practitionerRoleHistorySorted.find(
-    (it) =>
-      it?.meta?.lastUpdated &&
-      lastModified &&
-      it?.meta?.lastUpdated <= lastModified!
-  )
-
-  const targetCode = result?.code?.find((element) => {
-    return element.coding?.[0].system === 'http://opencrvs.org/specs/types'
-  })
-
-  const role = targetCode?.coding?.[0].code
-  const systemRole = result?.code?.[0].coding?.[0].code
-
-  return { role, systemRole }
 }
